+++ crates/ruma-events/src/room/encrypted.rs
@@ -5,7 +5,7 @@
 use std::{borrow::Cow, collections::BTreeMap};
 
 use js_int::UInt;
-use ruma_common::{serde::JsonObject, EventId, OwnedDeviceId, OwnedEventId};
+use ruma_common::{serde::JsonObject, OwnedDeviceId, OwnedEventId};
 use ruma_macros::EventContent;
 use serde::{Deserialize, Serialize};
 
@@ -118,36 +118,21 @@ impl Relation {
             Relation::Reference(_) => Some(RelationType::Reference),
             Relation::Annotation(_) => Some(RelationType::Annotation),
             Relation::Thread(_) => Some(RelationType::Thread),
-            Relation::_Custom(c) => Some(c.rel_type.as_str().into()),
-        }
-    }
-
-    /// The ID of the event this relates to.
-    ///
-    /// This is the `event_id` field at the root of an `m.relates_to` object, except in the case of
-    /// a reply relation where it's the `event_id` field in the `m.in_reply_to` object.
-    pub fn event_id(&self) -> &EventId {
-        match self {
-            Relation::Reply { in_reply_to } => &in_reply_to.event_id,
-            Relation::Replacement(r) => &r.event_id,
-            Relation::Reference(r) => &r.event_id,
-            Relation::Annotation(a) => &a.event_id,
-            Relation::Thread(t) => &t.event_id,
-            Relation::_Custom(c) => &c.event_id,
+            Relation::_Custom(c) => c.rel_type(),
         }
     }
 
     /// The associated data.
     ///
-    /// The returned JSON object won't contain the `rel_type` field, use
-    /// [`.rel_type()`][Self::rel_type] to access it. It also won't contain data
-    /// outside of `m.relates_to` (e.g. `m.new_content` for `m.replace` relations).
+    /// The returned JSON object holds the contents of `m.relates_to`, including `rel_type` and
+    /// `event_id` if present, but not things like `m.new_content` for `m.replace` relations that
+    /// live next to `m.relates_to`.
     ///
     /// Prefer to use the public variants of `Relation` where possible; this method is meant to
     /// be used for custom relations only.
     pub fn data(&self) -> Cow<'_, JsonObject> {
-        if let Relation::_Custom(c) = self {
-            Cow::Borrowed(&c.data)
+        if let Relation::_Custom(CustomRelation(data)) = self {
+            Cow::Borrowed(data)
         } else {
             Cow::Owned(self.serialize_data())
         }