+++ crates/ruma-events/CHANGELOG.md
@@ -1,5 +1,15 @@
 # [unreleased]
 
+Breaking changes:
+
+- Remove `event_id` methods from relation types
+
+Improvements:
+
+- Don't fail event content parsing on invalid relation
+  - We previously already accepted custom or slightly malformed relations
+  - Now, even invalid / missing `rel_type` and `event_id` are accepted
+
 # 0.27.9
 
 Bug fixes: